+++ src/components/UI/Records/features/TransactionFormFields/CurrencyAdornment.tsx
@@ -1,7 +1,7 @@
 import { InputAdornment } from '../../../../../styles';
 import { TypeOfRecord } from '../../../../../globalInterface';
 
-const CurrencyAdornment = ({ typeOfRecord }: { typeOfRecord: TypeOfRecord }) => {
+const CurrencyAdornment = ({ typeOfRecord }: { typeOfRecord?: TypeOfRecord }) => {
   const isExpense = typeOfRecord === 'expense';
   const isIncome = typeOfRecord === 'income';
   return (